% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtree.R
\name{qtree}
\alias{qtree}
\title{Scalable Qauntile Regression Trees}
\usage{
qtree(formula, data, mindev = 0.01, mincut = 5, minsize = 10, tau = 0.5,
  na.action = na.pass, model = FALSE, x = FALSE, y = TRUE, wts = TRUE,
  ...)
}
\value{
A class outtree
}
\description{
qtree will calculate quantile regression trees
}
